requirements.txt
streamlit>=1.35.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
# Google OAuth authentication
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
# Cookie-based session persistence
extra-streamlit-components>=0.1.71
